test: Add missing tests for n_permute_r
- Create `Tests/test_permutation.py` specifically for `n_permute_r`.
- Move the `TestPermutation` class out of `Tests/test_DiscreteMath.py` into the newly created dedicated file.
- Improve strictness of `test_n_permute_r_invalid` by using `assertRaisesRegex` with the expected error message string.
- Clean up the imports from `test_DiscreteMath.py` related to `n_permute_r`.
